Refactor disintegration debug controller
